pyproject.toml
[project]
name = "add-border"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pillow>=12.2.0",
]
